Add snarkvm-level module access
