+++ Cargo.toml
@@ -21,46 +21,123 @@ edition = "2018"
 members = [
   "algorithms",
   "curves",
-  "derives",
+  "derives", # excluded from [features], for internal use primarily
   "dpc",
   "fields",
   "gadgets",
   "ledger",
-  "marlin",
+  "marlin", # excluded from [features], to be merged in `algorithms` and `gadgets` once stable
   "parameters",
-  "polycommit",
-  "profiler",
+  "polycommit", # excluded from [features], to be merged in `algorithms` and `gadgets` once stable
+  "profiler", # excluded from [features], for internal use primarily
   "r1cs",
   "utilities",
-  "wasm",
-  ".integration"
+  "wasm", # excluded from [features], WASM contexts will typically import this crate directly
+  ".integration" # excluded from [features], for CI only
 ]
+
+[lib]
+path = "snarkvm/lib.rs"
 
 [[bin]]
 name = "snarkvm"
-path = "snarkvm/main.rs"
+path = "snarkvm/cli/main.rs"
+required-features = ["cli"]
+
+[features]
+default = ["core"]
+#
+# `core` is defined as the commonly-used members of snarkVM with default features.
+#
+# If you need more granular control of members and/or member features,
+# please import the requisite crates directly and enable features as needed.
+core = ["dpc", "ledger", "utilities"]
+full = ["algorithms", "curves", "dpc", "fields", "gadgets", "ledger", "parameters", "r1cs", "utilities"]
+
+## snarkVM workspace ##
+algorithms = ["snarkvm-algorithms"]
+curves = ["snarkvm-curves"]
+dpc = ["snarkvm-dpc"]
+fields = ["snarkvm-fields"]
+gadgets = ["snarkvm-gadgets"]
+ledger = ["snarkvm-ledger"]
+parameters = ["snarkvm-parameters"]
+r1cs = ["snarkvm-r1cs"]
+utilities = ["snarkvm-utilities"]
+
+## snarkVM CLI ##
+cli = ["anyhow", "colored", "self_update", "structopt", "thiserror"]
+
+## CI ##
+noconfig = []
+
+[dependencies.snarkvm-algorithms]
+path = "./algorithms"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-curves]
+path = "./curves"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-dpc]
+path = "./dpc"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-fields]
+path = "./fields"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-gadgets]
+path = "./gadgets"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-ledger]
+path = "./ledger"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-parameters]
+path = "./parameters"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-r1cs]
+path = "./r1cs"
+version = "0.7.5"
+optional = true
+
+[dependencies.snarkvm-utilities]
+path = "./utilities"
+version = "0.7.5"
+optional = true
 
 [dependencies.anyhow]
 version = "1.0.42"
+optional = true
 
 [dependencies.colored]
 version = "2"
+optional = true
 
 [dependencies.self_update]
 version = "0.27"
+optional = true
 
 [dependencies.structopt]
 version = "0.3"
+optional = true
 
 [dependencies.thiserror]
 version = "1.0"
+optional = true
 
 [dev-dependencies.rusty-hook]
 version = "0.11.2"
-
-[features]
-default = [ ]
-noconfig = [ ]
 
 [profile.release]
 opt-level = 3
