+++ ledger/src/ledger/tests.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::prelude::*;
+use crate::{ledger::*, prelude::*};
 use snarkvm_dpc::parameters::testnet2::Testnet2Parameters;
 
 #[test]
