+++ ledger/src/lib.rs
@@ -64,7 +64,6 @@ pub mod errors;
 pub use errors::*;
 
 pub mod ledger;
-pub use ledger::*;
 
 pub mod posw;
 
@@ -72,5 +71,5 @@ pub mod traits;
 pub use traits::*;
 
 pub mod prelude {
-    pub use crate::{block::*, errors::*, ledger::*, traits::*};
+    pub use crate::{block::*, errors::*, traits::*};
 }
