+++ snarkvm/cli/cli.rs
