+++ snarkvm/cli/errors.rs
